runtime compiles on Windows
