+++ runtime/spec_linux.go
@@ -2,4 +2,5 @@ package runtime
 
 import "github.com/opencontainers/specs"
 
-type platformSpec specs.LinuxSpec
+type PlatformSpec specs.LinuxSpec
+type ProcessSpec specs.Process
