+++ supervisor/sort_test.go
@@ -6,7 +6,6 @@ import (
 	"testing"
 
 	"github.com/docker/containerd/runtime"
-	"github.com/opencontainers/specs"
 )
 
 type testProcess struct {
@@ -45,8 +44,8 @@ func (p *testProcess) Container() runtime.Container {
 	return nil
 }
 
-func (p *testProcess) Spec() specs.Process {
-	return specs.Process{}
+func (p *testProcess) Spec() runtime.ProcessSpec {
+	return runtime.ProcessSpec{}
 }
 
 func (p *testProcess) Signal(os.Signal) error {
